+++ lucadeal/test/invoice_test.rb
@@ -6,8 +6,8 @@ class LucaDeal::InvoiceTest < Minitest::Test
   include LucaRecord::IO
 
   def setup
-    FileUtils.chdir(LucaSupport::Config::Pjdir)
-    LucaDeal::Setup.create_project(LucaSupport::Config::Pjdir)
+    FileUtils.chdir(LucaSupport::PJDIR)
+    LucaDeal::Setup.create_project(LucaSupport::PJDIR)
   end
 
   def teardown