+++ lucarecord/lib/luca_record/base.rb
@@ -6,6 +6,8 @@ require 'luca_support'
 
 module LucaRecord
   class Base
+    CONFIG = LucaSupport::CONFIG
+    PJDIR = LucaSupport::PJDIR
     include LucaRecord::IO
     include LucaSupport::View
   end